Switch sim_time and sim_time_multiplier from f32 to f64
Prevents precision loss during long simulations. Physics calculations in
CelestialBody::update cast to f32 internally since glam uses f32 throughout.
Also fixes a pre-existing compile error in the shader-watcher (event.paths → event.path),
and updates CLAUDE.md to reflect the current project state.

+++ src/scene.rs
@@ -59,8 +59,7 @@ impl Scene {
         }
     }
 
-    /// TODO: change from Duration to sim_time
-    pub fn update(&mut self, sim_time: f32, queue: &wgpu::Queue) {
+    pub fn update(&mut self, sim_time: f64, queue: &wgpu::Queue) {
         // update celestial bodies (including their local transforms)
         for cb in &mut self.celestial_bodies {
             cb.update(sim_time);
